+++ docsrc/conf.py
@@ -22,12 +22,12 @@ import merrycrispr
 
 # -- Project information -----------------------------------------------------
 
-project = 'MerryCRISPR'
-author = 'Miles Smith'
-copyright = f'{datetime.now():%Y}, {author}.'
+project = "MerryCRISPR"
+author = "Miles Smith"
+copyright = f"{datetime.now():%Y}, {author}."
 
 # The full version, including alpha/beta/rc tags
-release = '2.0'
+release = "2.0"
 
 
 # -- General configuration ---------------------------------------------------
@@ -36,38 +36,38 @@ release = '2.0'
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
+    "sphinx.ext.autodoc",
     # 'sphinx.ext.intersphinx',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.autosummary',
-    'sphinx_autodoc_typehints',
-    'sphinx.ext.autosectionlabel',
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autosummary",
+    "sphinx_autodoc_typehints",
+    "sphinx.ext.autosectionlabel",
 ]
 
 # Generate the API documentation when building
 autosummary_generate = True
-autodoc_member_order = 'bysource'
+autodoc_member_order = "bysource"
 # autodoc_default_flags = ['members']
 napoleon_google_docstring = False
 napoleon_numpy_docstring = True
 napoleon_include_init_with_doc = False
 napoleon_use_rtype = True  # having a separate entry generally helps readability
 napoleon_use_param = True
-napoleon_custom_sections = [('Params', 'Parameters')]
+napoleon_custom_sections = [("Params", "Parameters")]
 todo_include_todos = False
 nitpicky = True  # Report broken links
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-source_suffix = '.rst'
-master_doc = 'index'
-default_role = 'literal'
-pygments_style = 'sphinx'
+templates_path = ["_templates"]
+source_suffix = ".rst"
+master_doc = "index"
+default_role = "literal"
+pygments_style = "sphinx"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', '.gitignore', '.ipynb']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store", ".gitignore", ".ipynb"]
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -75,16 +75,14 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', '.gitignore', '.ipynb']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
-html_theme_options = dict(
-    navigation_depth=2,
-)
-html_static_path = ['_static']
+html_theme = "sphinx_rtd_theme"
+html_theme_options = dict(navigation_depth=2)
+html_static_path = ["_static"]
 html_show_sphinx = False
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 autosectionlabel_prefix_document = True